Add RecordManager and modify close button problem
